+++ packages/contracts/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Removed the `SignatureValidatorSet` event from `IDAO`.
 - Removed unused `ERC1271Mock` contract.
 - Deprecated the `setSignatureValidator` function and `signatureValidator` variable in `DAO`.