+++ packages/contracts/CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed the `SignatureValidatorSet` event from `IDAO`.
 - Removed unused `ERC1271Mock` contract.
-- Deprecated the `setSignatureValidator` function and `signatureValidator` variable in `DAO`.
+- Removed the `setSignatureValidator` function and `signatureValidator` variable in `DAO`. In places, where the function must remain to not alter the `IDAO` interface ID, it will revert and explanatory notes are put in place..
 
 ## v1.3.0-rc0